+++ packages/stream_chat_persistence/lib/src/mapper/pinned_message_mapper.dart
@@ -13,6 +13,7 @@ extension PinnedMessageEntityX on PinnedMessageEntity {
     List<Reaction>? ownReactions,
     Message? quotedMessage,
     Poll? poll,
+    Draft? draft,
   }) =>
       Message(
         shadowed: shadowed,
@@ -52,6 +53,7 @@ extension PinnedMessageEntityX on PinnedMessageEntity {
             mentionedUsers.map((e) => User.fromJson(jsonDecode(e))).toList(),
         i18n: i18n,
         restrictedVisibility: restrictedVisibility,
+        draft: draft,
       );
 }